Split out functions, will make into classes next
