+++ server.js
@@ -12,8 +12,10 @@ const options = {
   key: key,
   cert: cert
 };
-const apiKey0 = require('./requestVarFile.js');
-const dbConfig = require('./dbConfig.js');
+
+require('./InitaliseMongoCollections.js');
+const recentLocations = require('./recentLocations.js');
+const newLocationSearch = require('./newLocationSearch.js');
 
 /*
 Express Server Creation/Parsing Tools
@@ -28,142 +30,12 @@ server.listen(port, function(){
 });
 
 /*
-MongoDB Connectors/Collection creation
-*/
-
-const MongoClient = require('mongodb').MongoClient;
-
-MongoClient.connect(dbConfig.url, function(err, db) {
-    if (err) throw new Error(err)
-    db.close();
-});
-
-MongoClient.connect(dbConfig.url, function(err, db) {
-    if (err) throw new Error(err)
-    let dbo = db.db(dbConfig.dbName);
-    dbo.createCollection("recentlocations", function(err, res) {
-        db.close();
-    });
-});
-
-/*
-External Web Request Methods
-*/
-
-function reusableRequest(url, method){
-    let options = {
-      'method': method,
-      'url': url,
-      'headers': {
-        headers : ''
-      }
-    };
-    return new Promise (function (resolve) {
-      request(options, function (error, response) {
-        if (error) throw new Error(error);
-        resolve(response.body);
-      });
-    });
-};
-
-/*
-Google Places API Request/Parsing 
-*/
-
-function newRequest(randomPlace){
-    return new Promise (async function (resolve) {
-        let getPlaceDetailsUrl = 'https://maps.googleapis.com/maps/api/place/details/json?place_id='+randomPlace.place_id+'&key='+apiKey0;
-        let placeDetailsJson = await reusableRequest(getPlaceDetailsUrl, 'GET');
-        let placeDetailsObj = JSON.parse(placeDetailsJson);
-        collectResults(placeDetailsObj);
-        jsonResponse = JSON.stringify(placeDetailsObj);
-        resolve(jsonResponse);
-    });
-};
-
-async function getResults(userCoordinates) {
-    let placesJson = await reusableRequest('https://maps.googleapis.com/maps/api/place/nearbysearch/json?key=' + apiKey0 + '&location=' + userCoordinates + '&rankby=distance&keyword =food&type=restaurant', 'GET');
-    let placesObj = JSON.parse(placesJson);
-
-    if (placesObj.status !== 'OK') {
-        throw new Error(placesObj.status);
-    };
-
-    let randomPlace = placesObj.results[ Math.floor(Math.random() * placesObj.results.length)];
-    let isCached = await cachedRequestCheck(randomPlace);
-    return new Promise (async function (resolve) {
-        if (isCached === true){
-            MongoClient.connect(dbConfig.url, function(err, db) {
-                if (err) throw new Error(err);
-                let dbo = db.db(dbConfig.dbName);
-                let query = {"result.place_id": randomPlace.place_id};
-                    dbo.collection("recentlocations").find(query).toArray(function(err, result) {
-                        if (err) throw new Error(err);
-                        db.close();
-                        resolve(result[0]);
-                    });
-                });
-        }else{
-            let result = await newRequest(randomPlace);
-            resolve(result);
-        }
-    })
-};
-
-function cachedRequestCheck(randomPlace){
-    return new Promise (function (resolve) {
-        MongoClient.connect(dbConfig.url, function(err, db) {
-            if (err) throw new Error(err);
-            let dbo = db.db(dbConfig.dbName);
-            let query = {"result.place_id": randomPlace.place_id};
-            dbo.collection("recentlocations").find(query).toArray(function(err, result) {
-                if (err) throw new Error(err);
-                db.close();
-                    if (result.length !== 0){
-                        isCached = true;
-                        resolve(isCached);
-                    }else{
-                        isCached = false;
-                        resolve(isCached);
-                    }
-            });
-        });
-    });
-};
-
-function collectResults(placeDetailsObj){
-    MongoClient.connect(dbConfig.url, function(err, db) {
-        if (err) throw new Error(err);
-        let dbo = db.db(dbConfig.dbName);
-        dbo.collection("recentlocations").insertOne(placeDetailsObj, function(err, res) {
-            if (err) throw new Error(err);
-            db.close();
-        });
-    });
-}
-
-function getRecentLocations(){
-    return new Promise (function (resolve) {
-        MongoClient.connect(dbConfig.url, function(err, db) {
-            if (err) throw new Error(err);
-            let dbo = db.db(dbConfig.dbName);
-            let mysort = { name: -1 };
-            dbo.collection("recentlocations").find().limit(3).sort(mysort).toArray(function(err, result) {
-                if (err) throw new Error(err);
-                db.close();
-              resolve(result);
-            });
-        });
-    });
-};
-
-/*
 API HANDLERS
 */
 
 app.post('/newlocationsearch', async (req, res) => {
     try{
-        let responseObject = await getResults(req.query.coordinates);
+        let responseObject = await newLocationSearch.getResults(req.query.coordinates);
         res.send(responseObject);
     }catch(error){
         console.error(error);
@@ -174,7 +46,7 @@ app.post('/newlocationsearch', async (req, res) => {
 
 app.get('/recentlocations', async (req, res) => {
     try{
-        let result = await getRecentLocations();
+        let result = await recentLocations.getRecentLocations();
         jsonResponse = JSON.stringify(result);
         res.send(jsonResponse);
     }catch(error){
@@ -185,6 +57,12 @@ app.get('/recentlocations', async (req, res) => {
 });
 
 app.get('/images', (req, res) => {
-    let url = `https://maps.googleapis.com/maps/api/place/photo?maxwidth=2000&photoreference=${req.query.photo_reference}&key=${apiKey0}`;
-    request.get(url).pipe(res);
+    try{
+        let url = `https://maps.googleapis.com/maps/api/place/photo?maxwidth=2000&photoreference=${req.query.photo_reference}&key=${apiKey0}`;
+        request.get(url).pipe(res);
+    }catch(error){
+        console.error(error);
+        res.status(500);
+        res.json({'message': 'Something went wrong, please try again.'});
+    };
 });
